Refoctor Map Component and Use configure AXIOS for API calls
