+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "antd": "^3.9.3",
+    "axios": "^0.18.0",
     "classnames": "^2.2.6",
     "leaflet": "^1.3.4",
     "leaflet-draw": "^1.0.3",
