+++ src/common/store/rootReducer.js
@@ -2,6 +2,6 @@
  * Root Reducer for the application
  */
 import { combineReducers } from 'redux';
-import { alerts, alert } from '../../map/reducer';
+import * as reducers from '../../map/reducer';
 
-export default combineReducers({ alerts, alert });
+export default combineReducers({ ...reducers });
