+++ src/map/actions.js
@@ -3,6 +3,7 @@ export const ALERT_GET_START = 'ALERT_GET_START';
 export const ALERTS_STORE = 'ALERTS_STORE';
 export const ALERT_STORE = 'ALERT_STORE';
 export const MAP_SET_POSITION = 'MAP_SET_POSITION';
+export const FILTER_SET_SEVERITY = 'FILTER_SET_SEVERITY';
 
 export const alertsGetStart = () => ({ type: ALERTS_GET_START });
 
@@ -24,4 +25,9 @@ export const alertGetStart = (id = null) => ({
   payload: {
     data: id,
   },
+});
+
+export const setSeverityFilter = severity => ({
+  type: FILTER_SET_SEVERITY,
+  payload: { severity },
 });
