+++ src/map/components/AlertLegend/alertlegend.css
@@ -1,0 +1,8 @@
+.AlertLegend {
+    display: flex;
+    position: absolute;
+    left: 70vw;
+    top: 90vh;
+    opacity: 0.7;
+    z-index: 1000;  
+}
